backend/ai-service/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain==0.3.17
google-generativeai==0.8.4
slowapi==0.1.9
requests==2.32.3
beautifulsoup4==4.13.3
pandas==2.2.3
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
redis==5.3.0
apscheduler==3.10.4
chromadb==0.4.24
